MODIFIED: Exception Handling To All APIs

+++ src/main/java/com/infogalaxy/librarymanagementsystemproject/service/AuthorService.java
@@ -1,12 +1,14 @@
 package com.infogalaxy.librarymanagementsystemproject.service;
 
 import com.infogalaxy.librarymanagementsystemproject.entity.AuthorEntity;
+import com.infogalaxy.librarymanagementsystemproject.exceptions.MemberNotFoundException;
 import com.infogalaxy.librarymanagementsystemproject.repo.IAuthorRepo;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class AuthorService implements IAuthorService{
@@ -27,22 +29,34 @@ public class AuthorService implements IAuthorService{
     }
 
     @Override
-    public AuthorEntity retrieveAuthorById(int id) {
-        AuthorEntity authorEntity = iAuthorRepo.findById(id).get() ;
-        return authorEntity;
+    public Optional<AuthorEntity> retrieveAuthorById(int id) {
+        Optional<AuthorEntity> authorEntity = iAuthorRepo.findById(id);
+        if (authorEntity.isPresent()) {
+            return authorEntity;
+        } else {
+            throw new MemberNotFoundException("Author with Given Id is not Found I Database");
+        }
     }
 
     @Override
     public AuthorEntity updateAuthorById(int id, AuthorEntity authorEntity) {
-        AuthorEntity authorEntityNew = iAuthorRepo.findById(id).get();
-        BeanUtils.copyProperties(authorEntity,authorEntityNew);
-        return iAuthorRepo.save(authorEntityNew);
+        Optional<AuthorEntity> authorEntityNew = iAuthorRepo.findById(id);
+        if (authorEntityNew.isPresent()) {
+            BeanUtils.copyProperties(authorEntity, authorEntityNew);
+            return iAuthorRepo.save(authorEntity);
+        } else {
+            throw new MemberNotFoundException("Author with Given Id is not Found I Database");
+        }
     }
 
     @Override
     public String deleteAuthorById(int id) {
-        AuthorEntity authorEntity = iAuthorRepo.findById(id).get();
-        iAuthorRepo.deleteById(id);
-        return "Author Deleted Successfully...";
+        Optional<AuthorEntity> authorEntity = iAuthorRepo.findById(id);
+        if (authorEntity.isPresent()) {
+            iAuthorRepo.deleteById(id);
+            return "Author Deleted Successfully...";
+        } else {
+            throw new MemberNotFoundException("Author with Given Id is not Found I Database");
+        }
     }
 }
