+++ src/main/java/com/infogalaxy/librarymanagementsystemproject/service/BookService.java
@@ -1,12 +1,14 @@
 package com.infogalaxy.librarymanagementsystemproject.service;
 
 import com.infogalaxy.librarymanagementsystemproject.entity.BookEntity;
+import com.infogalaxy.librarymanagementsystemproject.exceptions.MemberNotFoundException;
 import com.infogalaxy.librarymanagementsystemproject.repo.IBookRepo;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BookService implements IBookService{
@@ -27,22 +29,35 @@ public class BookService implements IBookService{
     }
 
     @Override
-    public BookEntity retrieveBookById(int id) {
-        BookEntity bookEntity = iBookRepo.findById(id).get();
-        return bookEntity;
+    public Optional<BookEntity> retrieveBookById(int id) {
+        Optional<BookEntity> bookEntity = iBookRepo.findById(id);
+        if (bookEntity.isPresent()){
+            return bookEntity;
+        } else  {
+            throw new MemberNotFoundException("Book with Given Id is not Found in Database");
+        }
+
     }
 
     @Override
     public BookEntity updateBookById(int id, BookEntity bookEntity) {
-        BookEntity bookEntityNew = iBookRepo.findById(id).get();
-        BeanUtils.copyProperties(bookEntity,bookEntityNew);
-        return iBookRepo.save(bookEntityNew);
+        Optional<BookEntity> bookEntityNew = iBookRepo.findById(id);
+        if (bookEntityNew.isPresent()) {
+            BeanUtils.copyProperties(bookEntity, bookEntityNew);
+            return iBookRepo.save(bookEntity);
+        } else {
+            throw new MemberNotFoundException("Book with Given Id is not Found in Database");
+        }
     }
 
     @Override
     public String deleteBookById(int id) {
-        BookEntity bookEntity = iBookRepo.findById(id).get();
-        iBookRepo.deleteById(id);
-        return "Book Deleted Successfully...";
+        Optional<BookEntity> bookEntity = iBookRepo.findById(id);
+        if (bookEntity.isPresent()) {
+            iBookRepo.deleteById(id);
+            return "Book Deleted Successfully...";
+        } else  {
+            throw new MemberNotFoundException("Book with Given Id is not Found in Database");
+        }
     }
 }
